+++ resources/lib/pytube/parser.py
@@ -127,11 +127,14 @@ def parse_for_object_from_startpoint(html, start_point):
     full_obj = find_object_from_startpoint(html, start_point)
     try:
         return json.loads(full_obj)
-    except ValueError:
-        try:
-            return ast.literal_eval(full_obj)
-        except (ValueError, SyntaxError):
-            raise HTMLParseError('Could not parse object.')
+    except Exception as e:
+        if type(e).__name__ == 'JSONDecodeError':
+            try:
+                return ast.literal_eval(full_obj)
+            except (ValueError, SyntaxError):
+                raise HTMLParseError('Could not parse object.')
+        else:
+            raise(e)
 
 
 def throttling_array_split(js_array):